# Fix `execute` deoptimisation caused by GraphQLEnumType