@leebyron 2018-03-29
Nice find. This was lazy to avoid computation and heap space during Schema construction since these enum lookup maps might not be necessary if enum values are never actually requested. Though that may have been a premature optimization. We'll try it out.

A way to fix this while retaining laziness would have been to just assign `undefined` to those properties during construction so the class shape remains the same.